split data stores
