+++ chat-core/src/DatabaseConnection.ts
@@ -1,13 +1,3 @@
-import { strict as assert } from "assert";
-import { MongoClient, Filter } from "mongodb";
-import { PageStore, Page, PersistedPage } from "./Page";
-import {
-  EmbeddedContentStore,
-  EmbeddedContent,
-  FindNearestNeighborsOptions,
-  WithScore,
-} from "./EmbeddedContent";
-
 export type DatabaseConnection = {
   /**
     Close the connection.
@@ -17,180 +7,3 @@ export type DatabaseConnection = {
   close(force?: boolean): Promise<void>;
   drop(): Promise<void>;
 };
-
-export interface MakeDatabaseConnectionParams {
-  connectionUri: string;
-  databaseName: string;
-}
-
-/**
-  OSS_TODO: in my opinion, we should do some slight refactors here.
-  i find it confusing that we are calling this makeDatabaseConnection,
-  when its core purpose is working with the PageStore and the EmbeddedContentStore.
-  My intuition is to split up `PageStore` and `EmbeddedContentStore` into separate
-  constructor funcs (i.e. have separate `makePageStore()` and `makeEmbeddedContentStore()`).
-  Then we can also remove the concept of a `DatabaseConnection`, which i don't think
-  is a particularly relevant abstraction for our project.
- */
-/**
-  Create a connection to the database.
- */
-export const makeDatabaseConnection = ({
-  connectionUri,
-  databaseName,
-}: MakeDatabaseConnectionParams): DatabaseConnection &
-  PageStore &
-  EmbeddedContentStore => {
-  const client = new MongoClient(connectionUri, {
-    serverSelectionTimeoutMS: 30000,
-  });
-  const db = client.db(databaseName);
-  const embeddedContentCollection =
-    db.collection<EmbeddedContent>("embedded_content");
-  const pagesCollection = db.collection<PersistedPage>("pages");
-  const instance: DatabaseConnection & PageStore & EmbeddedContentStore = {
-    async drop() {
-      await db.dropDatabase();
-    },
-
-    async close(force) {
-      client.close(force);
-    },
-
-    async loadEmbeddedContent({ page }) {
-      return await embeddedContentCollection.find(pageIdentity(page)).toArray();
-    },
-
-    async deleteEmbeddedContent({ page }) {
-      const deleteResult = await embeddedContentCollection.deleteMany(
-        pageIdentity(page)
-      );
-      if (!deleteResult.acknowledged) {
-        throw new Error("EmbeddedContent deletion not acknowledged!");
-      }
-    },
-
-    async updateEmbeddedContent({ page, embeddedContent }) {
-      assert(embeddedContent.length !== 0);
-      embeddedContent.forEach((embeddedContent) => {
-        assert(
-          embeddedContent.sourceName === page.sourceName &&
-            embeddedContent.url === page.url,
-          `EmbeddedContent source/url (${embeddedContent.sourceName} / ${embeddedContent.url}) must match give page source/url (${page.sourceName} / ${page.url})!`
-        );
-      });
-      await client.withSession(async (session) => {
-        await session.withTransaction(async () => {
-          // First delete all the embeddedContent for the given page
-          const deleteResult = await embeddedContentCollection.deleteMany(
-            pageIdentity(page),
-            { session }
-          );
-          if (!deleteResult.acknowledged) {
-            throw new Error("EmbeddedContent deletion not acknowledged!");
-          }
-
-          // Insert the embedded content for the page
-          const insertResult = await embeddedContentCollection.insertMany(
-            [...embeddedContent],
-            {
-              session,
-            }
-          );
-
-          if (!insertResult.acknowledged) {
-            throw new Error("EmbeddedContent insertion not acknowledged!");
-          }
-          const { insertedCount } = insertResult;
-          if (insertedCount !== embeddedContent.length) {
-            throw new Error(
-              `Expected ${embeddedContent.length} inserted, got ${insertedCount}`
-            );
-          }
-        });
-      });
-    },
-
-    async findNearestNeighbors(vector, options) {
-      const {
-        indexName,
-        path,
-        k,
-        minScore,
-        filter,
-      }: Partial<FindNearestNeighborsOptions> = {
-        // Default options
-        indexName: "default",
-        path: "embedding",
-        k: 3,
-        minScore: 0.9,
-
-        // User options override
-        ...(options ?? {}),
-      };
-      return embeddedContentCollection
-        .aggregate<WithScore<EmbeddedContent>>([
-          {
-            $search: {
-              index: indexName,
-              knnBeta: {
-                vector,
-                path,
-                k,
-                filter,
-              },
-            },
-          },
-          {
-            $addFields: {
-              score: {
-                $meta: "searchScore",
-              },
-            },
-          },
-          { $match: { score: { $gte: minScore } } },
-        ])
-        .toArray();
-    },
-
-    async loadPages(args) {
-      const filter: Filter<PersistedPage> = {};
-      if (args?.sources !== undefined) {
-        filter.sourceName = { $in: args.sources };
-      }
-      if (args?.updated !== undefined) {
-        filter.updated = { $gte: args.updated };
-      }
-      return pagesCollection.find(filter).toArray();
-    },
-
-    async updatePages(pages) {
-      await Promise.all(
-        pages.map(async (page) => {
-          const result = await pagesCollection.updateOne(
-            pageIdentity(page),
-            { $set: page },
-            { upsert: true }
-          );
-          if (!result.acknowledged) {
-            throw new Error(`update pages not acknowledged!`);
-          }
-          if (!result.modifiedCount && !result.upsertedCount) {
-            throw new Error(
-              `Page ${JSON.stringify(pageIdentity(page))} not updated!`
-            );
-          }
-        })
-      );
-    },
-  };
-  return instance;
-};
-
-/**
-  Returns a query filter that represents a unique page in the system.
- */
-export const pageIdentity = ({ url, sourceName }: Page) => ({
-  url,
-  sourceName,
-});
